+++ GUI/Controls/ManageMods.cs
@@ -335,24 +335,27 @@ namespace CKAN.GUI
 
         private void labelMenuItem_Click(object? sender, EventArgs? e)
         {
-            if (user != null && manager != null && currentInstance != null && SelectedModule != null)
+            if (user != null
+                && manager != null
+                && currentInstance != null
+                && SelectedModule != null
+                && sender is ToolStripMenuItem item
+                && item.Tag is ModuleLabel mlbl)
             {
-                var item   = sender    as ToolStripMenuItem;
-                var mlbl   = item?.Tag as ModuleLabel;
-                if (item?.Checked ?? false)
+                if (item.Checked)
                 {
-                    mlbl?.Add(currentInstance.game, SelectedModule.Identifier);
+                    mlbl.Add(currentInstance.game, SelectedModule.Identifier);
                 }
                 else
                 {
-                    mlbl?.Remove(currentInstance.game, SelectedModule.Identifier);
+                    mlbl.Remove(currentInstance.game, SelectedModule.Identifier);
                 }
                 var registry = RegistryManager.Instance(currentInstance, repoData).registry;
                 mainModList.ReapplyLabels(SelectedModule, Conflicts?.ContainsKey(SelectedModule) ?? false,
                                           currentInstance.Name, currentInstance.game, registry);
                 ModuleLabelList.ModuleLabels.Save(ModuleLabelList.DefaultPath);
                 UpdateHiddenTagsAndLabels();
-                if (mlbl?.HoldVersion ?? false)
+                if (mlbl.HoldVersion || mlbl.IgnoreMissingFiles)
                 {
                     UpdateCol.Visible = UpdateAllToolButton.Enabled =
                         mainModList.ResetHasUpdate(currentInstance, registry, ChangeSet, ModGrid.Rows);
@@ -538,19 +541,18 @@ namespace CKAN.GUI
         {
             WithFrozenChangeset(() =>
             {
-                foreach (var gmod in mainModList.full_list_of_mod_rows
-                                                .Values
-                                                .Select(row => row.Tag)
-                                                .OfType<GUIMod>())
+                var checkboxes = mainModList.full_list_of_mod_rows
+                                            .Values
+                                            .Where(row => row.Tag is GUIMod {Identifier: string ident}
+                                                          && (!Main.Instance?.LabelsHeld(ident) ?? false))
+                                            .SelectWithCatch(row => row.Cells[UpdateCol.Index],
+                                                             (row, exc) => null)
+                                            .OfType<DataGridViewCheckBoxCell>();
+                foreach (var checkbox in checkboxes)
                 {
-                    if (gmod?.HasUpdate ?? false)
-                    {
-                        if (!Main.Instance?.LabelsHeld(gmod.Identifier) ?? false)
-                        {
-                            gmod.SelectedMod = gmod.LatestCompatibleMod;
-                        }
-                    }
+                    checkbox.Value = true;
                 }
+                ModGrid.CommitEdit(DataGridViewDataErrorContexts.Commit);
 
                 // only sort by Update column if checkbox in settings checked
                 if (guiConfig?.AutoSortByUpdate ?? false)
@@ -935,7 +937,7 @@ namespace CKAN.GUI
                                             : gmod.LatestCompatibleMod
                                         : gmod.InstalledMod?.Module;
 
-                                    if (nowChecked && gmod.SelectedMod == gmod.LatestCompatibleMod)
+                                    if (gmod.SelectedMod == gmod.LatestCompatibleMod)
                                     {
                                         // Reinstall, force update without change
                                         UpdateChangeSetAndConflicts(currentInstance,