+++ GUI/Model/ModList.cs
@@ -470,14 +470,17 @@ namespace CKAN.GUI
                                          // Skip reinstalls
                                          .Where(upg => upg.Mod != upg.targetMod)
                                          .ToArray();
-                if (upgrades.Length > 0)
+                if (upgrades.Length > 0 && instance != null)
                 {
                     var upgradeable = registry.CheckUpgradeable(instance,
                                                                 // Hold identifiers not chosen for upgrading
                                                                 registry.Installed(false)
                                                                         .Select(kvp => kvp.Key)
                                                                         .Except(upgrades.Select(ch => ch.Mod.identifier))
-                                                                        .ToHashSet())
+                                                                        .ToHashSet(),
+                                                                ModuleLabelList.ModuleLabels
+                                                                               .IgnoreMissingIdentifiers(instance)
+                                                                               .ToHashSet())
                                               [true]
                                               .ToDictionary(m => m.identifier,
                                                             m => m);
@@ -523,6 +526,8 @@ namespace CKAN.GUI
         {
             var upgGroups = registry.CheckUpgradeable(inst,
                                                       ModuleLabelList.ModuleLabels.HeldIdentifiers(inst)
+                                                                                  .ToHashSet(),
+                                                      ModuleLabelList.ModuleLabels.IgnoreMissingIdentifiers(inst)
                                                                                   .ToHashSet());
             var dlls = registry.InstalledDlls.ToList();
             foreach ((var upgradeable, var mods) in upgGroups)
@@ -559,8 +564,13 @@ namespace CKAN.GUI
                         full_list_of_mod_rows[ident] =
                             MakeRow(gmod, ChangeSet, inst.Name, inst.game);
                     var rowIndex = row.Index;
+                    var selected = row.Selected;
                     rows.Remove(row);
                     rows.Insert(rowIndex, newRow);
+                    if (selected)
+                    {
+                        rows[rowIndex].Selected = true;
+                    }
                 }
             }
         }
@@ -583,14 +593,16 @@ namespace CKAN.GUI
                           config?.HideEpochs ?? false, config?.HideV ?? false);
 
         private static IEnumerable<GUIMod> GetGUIMods(IRegistryQuerier      registry,
-                                               RepositoryDataManager repoData,
-                                               GameInstance          inst,
-                                               GameVersionCriteria   versionCriteria,
-                                               HashSet<string>       installedIdents,
-                                               bool                  hideEpochs,
-                                               bool                  hideV)
+                                                      RepositoryDataManager repoData,
+                                                      GameInstance          inst,
+                                                      GameVersionCriteria   versionCriteria,
+                                                      HashSet<string>       installedIdents,
+                                                      bool                  hideEpochs,
+                                                      bool                  hideV)
             => registry.CheckUpgradeable(inst,
                                          ModuleLabelList.ModuleLabels.HeldIdentifiers(inst)
+                                                                     .ToHashSet(),
+                                         ModuleLabelList.ModuleLabels.IgnoreMissingIdentifiers(inst)
                                                                      .ToHashSet())
                        .SelectMany(kvp => kvp.Value
                                              .Select(mod => registry.IsAutodetected(mod.identifier)