+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## v1.35.1
 
+### Features
+
+- [Multiple] Improvements for missing file checks (#4213 by: HebaruSan)
+
 ### Bugfixes
 
 - [Core] Skip corrupted .acf files in Steam library (#4200 by: HebaruSan)